feat :: reminder

+++ main.go
@@ -32,7 +32,7 @@ func Ticker24Hour() {
 
 func init() {
 	db.Connect()
-	controller.MessageStopper = make(map[uint](chan bool))
+	controller.MessageStopper = make(map[uint64](chan bool))
 	controller.GetReminders()
 
 	now := time.Now()
